apps/agents/requirements.txt
boto3
botocore
gitdb
GitPython
jmespath
pyspellchecker
python-dateutil
s3transfer
six
smmap
urllib3
python-dotenv
bcrypt
pymongo